refactor: moded to the modals folder
